backend/requirements.txt
# Gyaan Saarthi Backend - FREE VERSION
# 100% Free Dependencies - No Paid APIs

# Django Core
Django==4.2.7
djangorestframework
django-cors-headers==4.3.0
django-filter==23.3
drf-spectacular==0.26.5

# Database
psycopg-binary
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# FREE AI - Ollama (Local LLM)
requests==2.31.0
httpx==0.26.0

# FREE Speech Recognition - Vosk
vosk==0.3.45
pydub==0.25.1
SpeechRecognition==3.10.1

# FREE TTS - pyttsx3 or gTTS
pyttsx3==2.90
gTTS==2.5.0

# Image Processing
Pillow
python-magic==0.4.27

# FREE File Storage - Cloudinary
cloudinary==1.38.0

# Task Queue (Optional - for background tasks)
celery==5.3.4
redis==5.0.1

# Utilities
python-decouple==3.8
python-multipart==0.0.9
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
ipython==8.20.0
black==24.1.1
setuptools
chromadb
sentence-transformers